requirements-api.txt
pytest
pytest
pytest-azurepipelines
selenium
Appium-Python-Client
requests

